progress on nf
