--- configurations/nf/exploreHomeConfiguration/data.ts
@@ -1,89 +1,0 @@
-import { SynapseConstants } from 'synapse-react-client'
-import { HomeExploreConfig } from '../../types/portal-config'
-import loadingScreen from '../loadingScreen'
-
-const sql = 'SELECT * FROM syn18488466'
-const unitDescription = 'Datasets'
-const synapseId = 'syn18488466'
-
-const rgbIndex = 0
-const facetName = 'tumorType'
-
-export const data: HomeExploreConfig = {
-  homePageSynapseObject: {
-    name: 'QueryWrapper',
-    props: {
-      rgbIndex,
-      facetName,
-      unitDescription,
-      name: 'Data',
-      initQueryRequest : {
-        concreteType: 'org.sagebionetworks.repo.model.table.QueryBundleRequest',
-        partMask: SynapseConstants.BUNDLE_MASK_QUERY_FACETS
-          | SynapseConstants.BUNDLE_MASK_QUERY_RESULTS,
-        query: {
-          sql,
-          isConsistent: false,
-          limit: 25,
-          offset: 0,
-        }
-      },
-    }
-  },
-  explorePageSynapseObject: {
-    name: 'QueryWrapperMenu',
-    props: {
-      loadingScreen,
-      rgbIndex,
-      type: SynapseConstants.CSBC_DATASET,
-      menuConfig: [
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetName: 'species',
-          facetAliases: {
-            species: 'Species',
-          }
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetName: 'Theme'
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetName: 'experimentalStrategy',
-          facetAliases: {
-            experimentalStrategy: 'Assay'
-          }
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetName: 'platform',
-          facetAliases: {
-            platform: 'Platform'
-          }
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetName: 'tumorType',
-          facetAliases: {
-            tumorType: 'Disease Type'
-          }
-        }
-      ],
-      facetName: 'tumorType',
-      facetAliases: {
-        tumorType: 'Disease Type',
-      },
-    },
-  }
-}
