--- configurations/nf/exploreHomeConfiguration/index.ts
@@ -1,15 +1,0 @@
-import datasets from './datasets'
-import files from './files'
-import publications from './publications'
-import studies from './studies'
-import tools from './tools'
-import funders from './funders'
-
-export {
-  studies,
-  datasets,
-  files,
-  publications,
-  tools,
-  funders,
-}
