+++ configurations/nf/homeExploreConfiguration/data.ts
@@ -2,27 +2,12 @@ import { SynapseConstants } from 'synapse-react-client'
 import { HomeExploreConfig } from '../../types/portal-config'
 import loadingScreen from '../loadingScreen'
 
-// tslint:disable-next-line:max-line-length
-const sql = 'SELECT grantType AS "Grant Type", centerName AS "Grant", consortium AS "Program", species AS "Species", fileFormat AS "Data Format", experimentalStrategy AS "Assay", platform AS "Platform", tumorType AS "Disease Type", sex AS "Gender", tissue AS "Tissue", name as "File Name"  FROM syn9630847'
+const sql = 'SELECT * FROM syn18488466'
+const unitDescription = 'Datasets'
+const synapseId = 'syn18488466'
 
-const facetAliases = {
-  grantType: 'Grant Type',
-  centerName: 'Grant',
-  consortium: 'Program',
-  species: 'Species',
-  fileFormat: 'Data Format',
-  experimentalStrategy: 'Assay',
-  platform: 'Platform',
-  tumorType: 'Disease Type',
-  sex: 'Gender',
-  tissue: 'Tissue',
-  name: 'File Name'
-}
-
-const rgbIndex = 6
-const unitDescription = 'files'
-const synapseId  = 'syn9630847'
-const facetName = 'grantType'
+const rgbIndex = 0
+const facetName = 'tumorType'
 
 export const data: HomeExploreConfig = {
   homePageSynapseObject: {
@@ -30,13 +15,10 @@ export const data: HomeExploreConfig = {
     props: {
       rgbIndex,
       facetName,
-      unitDescription,
       loadingScreen,
-      name: 'Files',
-      facetAliases: {
-        grantType: 'Grant Type',
-      },
-      initQueryRequest: {
+      unitDescription,
+      name: 'Data',
+      initQueryRequest : {
         concreteType: 'org.sagebionetworks.repo.model.table.QueryBundleRequest',
         partMask: SynapseConstants.BUNDLE_MASK_QUERY_FACETS
           | SynapseConstants.BUNDLE_MASK_QUERY_RESULTS,
@@ -46,104 +28,63 @@ export const data: HomeExploreConfig = {
           limit: 25,
           offset: 0,
         }
-      }
+      },
     }
   },
   explorePageSynapseObject: {
     name: 'QueryWrapperMenu',
     props: {
-      rgbIndex,
       loadingScreen,
+      rgbIndex,
+      type: SynapseConstants.CSBC_DATASET,
       menuConfig: [
         {
           sql,
           synapseId,
           unitDescription,
-          facetAliases,
-          title: 'Grant Type',
-          facetName: 'grantType',
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetAliases,
-          title: 'Grant',
-          facetName: 'centerName',
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetAliases,
-          title: 'Program',
-          facetName: 'consortium',
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetAliases,
-          title: 'Species',
           facetName: 'species',
+          facetAliases: {
+            species: 'Species',
+          }
         },
         {
           sql,
           synapseId,
           unitDescription,
-          facetAliases,
-          title: 'Theme',
           facetName: 'Theme'
         },
         {
           sql,
           synapseId,
           unitDescription,
-          facetAliases,
-          title: 'Data Format',
-          facetName: 'fileFormat',
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetAliases,
-          title: 'Assay',
           facetName: 'experimentalStrategy',
+          facetAliases: {
+            experimentalStrategy: 'Assay'
+          }
         },
         {
           sql,
           synapseId,
           unitDescription,
-          facetAliases,
-          title: 'Platform',
           facetName: 'platform',
+          facetAliases: {
+            platform: 'Platform'
+          }
         },
         {
           sql,
           synapseId,
           unitDescription,
-          facetAliases,
-          title: 'Disease Type',
           facetName: 'tumorType',
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetAliases,
-          title: 'Gender',
-          facetName: 'sex',
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetAliases,
-          title: 'Tissue',
-          facetName: 'tissue',
-        },
-      ]
-    }
+          facetAliases: {
+            tumorType: 'Disease Type'
+          }
+        }
+      ],
+      facetName: 'tumorType',
+      facetAliases: {
+        tumorType: 'Disease Type',
+      },
+    },
   }
 }
