+++ configurations/nf/homeExploreConfiguration/datasets.ts
@@ -1,87 +1,76 @@
 import { SynapseConstants } from 'synapse-react-client'
 import { HomeExploreConfig } from '../../types/portal-config'
 import loadingScreen from '../loadingScreen'
-export const datasetsSql = `SELECT * FROM syn18488466 WHERE ( ( "featured" = 'TRUE' ) )`
-const sql = 'SELECT * FROM syn18488466'
-const unitDescription = 'Datasets'
-const synapseId = 'syn18488466'
-const rgbIndex = 0
 
-export const datasets: HomeExploreConfig = {
+const sql = 'SELECT * FROM syn16859580'
+export const datasetsSql = sql
+const type = 'dataset'
+const unitDescription = 'datasets'
+const rgbIndex = 5
+const synapseId = 'syn16859580'
+
+const facetAliases = {
+  diseaseFocus: 'Disease Focus',
+  tumorType: 'Tumor Type',
+  fundingAgency: 'Funding Agency',
+}
+
+const datasets: HomeExploreConfig = {
   homePageSynapseObject: {
     name: 'QueryWrapper',
     props: {
       unitDescription,
+      rgbIndex,
       loadingScreen,
-      rgbIndex: 0,
-      facetName: 'tumorType',
       name: 'Datasets',
-      initQueryRequest : {
+      facetName: 'diseaseFocus',
+      initQueryRequest: {
         concreteType: 'org.sagebionetworks.repo.model.table.QueryBundleRequest',
-        partMask: SynapseConstants.BUNDLE_MASK_QUERY_FACETS
+        partMask:
+          SynapseConstants.BUNDLE_MASK_QUERY_COLUMN_MODELS
+          | SynapseConstants.BUNDLE_MASK_QUERY_FACETS
           | SynapseConstants.BUNDLE_MASK_QUERY_RESULTS,
         query: {
           sql,
           isConsistent: false,
           limit: 25,
           offset: 0,
-        }
-      },
-      facetAliases: {
-        tumorType: 'Disease Type',
-      },
+        },
+      }
     }
   },
   explorePageSynapseObject: {
     name: 'QueryWrapperMenu',
     props: {
+      unitDescription,
       rgbIndex,
+      type,
       loadingScreen,
-      type: SynapseConstants.CSBC_DATASET,
       menuConfig: [
         {
           sql,
-          synapseId,
+          facetAliases,
           unitDescription,
-          facetName: 'species',
-          facetAliases: {
-            species: 'Species',
-          }
+          synapseId,
+          facetName: 'diseaseFocus',
         },
         {
           sql,
-          synapseId,
+          facetAliases,
           unitDescription,
-          facetName: 'Theme'
-        },
-        {
-          sql,
           synapseId,
-          unitDescription,
-          facetName: 'experimentalStrategy',
-          facetAliases: {
-            experimentalStrategy: 'Assay'
-          }
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
-          facetName: 'platform',
-          facetAliases: {
-            platform: 'Platform'
-          }
-        },
-        {
-          sql,
-          synapseId,
-          unitDescription,
           facetName: 'tumorType',
-          facetAliases: {
-            tumorType: 'Disease Type'
-          }
-        }
-      ],
+        },
+        {
+          sql,
+          facetAliases,
+          unitDescription,
+          synapseId,
+          facetName: 'fundingAgency',
+        },
+      ]
     }
   }
 }
+
+export default datasets
