+++ configurations/nf/homeExploreConfiguration/funders.ts
