+++ configurations/nf/homeExploreConfiguration/index.ts
@@ -1,9 +1,15 @@
-import { data } from './data'
-import { datasets } from './datasets'
-import { grants } from './grants'
-import { publications } from './publications'
-import { studies } from './studies'
+import datasets from './datasets'
+import files from './files'
+import publications from './publications'
+import studies from './studies'
+import tools from './tools'
+import funders from './funders'
 
-export { data, datasets, grants, publications, studies }
-
-export default { data, datasets, grants, publications, studies }
+export {
+  studies,
+  datasets,
+  files,
+  publications,
+  tools,
+  funders,
+}
