+++ configurations/nf/homeExploreConfiguration/tools.ts
