+++ configurations/nf/loadingScreen.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { BarLoader } from 'react-spinners'
 
-const loadingBar = <div className="bar-loader"><BarLoader color="#47337D" loading={true}/></div>
+const loadingBar = <div className="bar-loader"><BarLoader color="#4DB7AD" loading={true}/></div>
 
 export default loadingBar
